backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
reportlab==4.2.5
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
resend==2.21.0
anthropic==0.39.0
stripe==14.2.0
cryptography>=41.0.0
PyYAML>=6.0
